requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy>=2.0.45,<2.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
transformers>=4.47.0
tokenizers>=0.21.0
torch>=2.9.0,<2.10
reportlab==4.2.0
python-multipart==0.0.9
email-validator==2.1.1
psycopg[binary]>=3.2.10,<3.3
python-dotenv==1.0.1
